Create new layout for blog page
This is the new layout for the blog page. I didn't go in and edit the files myself because it looks like a new layout file is required. (To make the background stretch to the width of the browser.) The only thing added here is everything in the BECCA ADDED THIS SECTION and the style tag in the header. You can setup the styles how you want. I'll upload the images as well. There may be a mobile change depending on what Wendell wants to do.